把resource csv 重构成了 prts data
